Renamed the AccountState test file to a more meaningful name. Abstracted a StateTestFixture class from the account state tests to simplify testing Jareel states
