+++ Zoo Game/Assets/Scripts/States/Editor/StateTestFixture.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 50dea627f80727241a7657e2ab14a9ea
+timeCreated: 1492975988
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
